+++ weekly_mission_1/example10/main.js
@@ -1,0 +1,18 @@
+const Pokemon = require('./pokemon')
+
+const pikachu = new Pokemon("pikachu")
+const bulbasaur = new Pokemon("bulbasaur")
+const squirtle = new Pokemon("squirtle")
+const charmander = new Pokemon("charmander")
+
+pikachu.sayHello()
+pikachu.sayMessage("Pikapi!!!!!")
+
+bulbasaur.sayHello()
+bulbasaur.sayMessage("Bulbasaur!!!")
+
+charmander.sayHello()
+charmander.sayMessage("Chaaar!!!")
+
+squirtle.sayHello()
+squirtle.sayMessage("Squirtle Squirtle Squirtle!!!")
