+++ weekly_mission_1/example10/pokemon.js
@@ -1,0 +1,15 @@
+class Pokemon{
+    constructor(name){
+        this.name = name;
+    }
+
+    sayHello(){
+        console.log(`Mi pokemon ${this.name} te saluda!!!`)
+    }
+
+    sayMessage(message){
+        console.log(`Mi pokemon ${this.name} dice: ${message}`)
+    }
+}
+
+module.exports = Pokemon;
