+++ weekly_mission_1/example9/main.js
@@ -1,18 +1,4 @@
-const Pokemon = require('./pokemon')
+import MyPokemon from './pokemon.js'
 
-const pikachu = new Pokemon("pikachu")
-const bulbasaur = new Pokemon("bulbasaur")
-const squirtle = new Pokemon("squirtle")
-const charmander = new Pokemon("charmander")
-
+const pikachu = new MyPokemon('Pikachu')
 pikachu.sayHello()
-pikachu.sayMessage("Pikapi!!!!!")
-
-bulbasaur.sayHello()
-bulbasaur.sayMessage("Bulbasaur!!!")
-
-charmander.sayHello()
-charmander.sayMessage("Chaaar!!!")
-
-squirtle.sayHello()
-squirtle.sayMessage("Squirtle Squirtle Squirtle!!!")
