+++ weekly_mission_1/example9/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "esm-syntax",
+  "version": "1.0.0",
+  "description": "",
+  "main": "main.js",
+  "type": "module",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
